requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
appdirs==1.4.4
argcomplete==1.10.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.8.2
bidict==0.22.1
black==22.1.0
bson==0.5.10
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.0.3
colorama==0.4.4
compressed-rtf==1.0.6
comtypes==1.1.14
configobj==5.0.8
configparser==5.3.0
debugpy==1.5.1
decorator==5.1.1
DeepFilterLib==0.4.0
deepfilternet==0.4.0
dnspython==2.3.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
email-validator==2.0.0.post2
entrypoints==0.4
etelemetry==0.3.0
exceptiongroup==1.1.2
executing==0.8.2
extract-msg==0.28.7
filelock==3.9.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.5
Flask-WTF==1.1.1
frozenlist==1.3.3
future==0.18.3
glob2==0.7
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
IMAPClient==2.1.0
ipykernel==6.8.0
ipython==8.0.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.1.2
jupyter-core==4.9.1
loguru==0.6.0
looseversion==1.0.3
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mongoengine==0.27.0
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.4
networkx==3.0
nibabel==5.0.1
nipype==1.8.5
numpy==1.23.1
olefile==0.46
openai==0.27.8
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pdfminer==20191125
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.4.1
prompt-toolkit==3.0.26
prov==2.0.0
pure-eval==0.2.2
PyAudio==0.2.13
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pydantic==2.3.0
pydantic_core==2.6.3
pydot==1.4.2
Pygments==2.11.2
PyJWT==2.7.0
pymongo==4.4.1
PyMuPDF==1.21.1
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF4==1.27.0
pypiwin32==223
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-pptx==0.6.21
python-socketio==5.8.0
pyttsx3==2.90
pytz==2022.7.1
pywin32==303
pyxnat==1.5
pyzmq==22.3.0
rdflib==6.2.0
redis==4.6.0
requests==2.28.2
rfc3986==1.5.0
scipy==1.10.0
selenium==4.10.0
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
sockets==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
spotipy==2.23.0
stack-data==0.1.4
textract==1.6.5
tk==0.1.0
tomli==2.0.0
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
traits==6.3.2
trio==0.22.1
trio-websocket==0.10.3
twilio==8.4.0
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wcwidth==0.2.5
webdriver-manager==3.8.6
Werkzeug==2.2.2
win32-setctime==1.1.0
wsproto==1.2.0
WTForms==3.0.1
xlrd==1.2.0
XlsxWriter==3.0.7
yarg==0.1.9
yarl==1.9.2
